Move Vault's customized pkcs7 fork into a shared location
 - The PKI plugin needs to use the customized pkcs7 fork
   so move it out from the aws credential plugin's package
   into a shared location
